moved merge code out of TGW and into Fate
Most of the code in the new DeleteRows and MergeTablets classes was
copied from TabletGroupWatcher with slight modifications (like adding
the Manager as parameter).

fixes #3650
